+++ src/main/kotlin/graphql/dataTransferObjects/DeletedAccount.kt
@@ -1,5 +1,5 @@
 package com.neelkamath.omniChatBackend.graphql.dataTransferObjects
 
-class DeletedAccount(private val userId: Int) : AccountsSubscription {
+class DeletedAccount(private val userId: Int) : AccountsSubscription, ChatAccountsSubscription {
     fun getUserId(): Int = userId
 }