+++ src/main/kotlin/graphql/dataTransferObjects/InvalidChatId.kt
@@ -5,6 +5,7 @@ import com.neelkamath.omniChatBackend.graphql.routing.Placeholder
 object InvalidChatId :
     ChatOnlineStatusesSubscription,
     ChatTypingStatusesSubscription,
+    ChatAccountsSubscription,
     CreateTextMessageResult,
     ForwardMessageResult,
     SearchChatMessagesResult,