+++ src/main/kotlin/graphql/engine/Operations.kt
@@ -73,8 +73,10 @@ private fun wireMutation(builder: TypeRuntimeWiring.Builder): TypeRuntimeWiring.
 private fun wireSubscription(builder: TypeRuntimeWiring.Builder): TypeRuntimeWiring.Builder = builder
     .dataFetcher("subscribeToMessages", ::subscribeToMessages)
     .dataFetcher("subscribeToChatMessages", ::subscribeToChatMessages)
+    .dataFetcher("subscribeToChatAccounts", ::subscribeToChatAccounts)
     .dataFetcher("subscribeToAccounts", ::subscribeToAccounts)
     .dataFetcher("subscribeToGroupChats", ::subscribeToGroupChats)
     .dataFetcher("subscribeToTypingStatuses", ::subscribeToTypingStatuses)
+    .dataFetcher("subscribeToChatTypingStatuses", ::subscribeToChatTypingStatuses)
     .dataFetcher("subscribeToOnlineStatuses", ::subscribeToOnlineStatuses)
     .dataFetcher("subscribeToChatOnlineStatuses", ::subscribeToChatOnlineStatuses)