+++ src/main/kotlin/graphql/routing/Subscriptions.kt
@@ -24,32 +24,36 @@ import org.reactivestreams.Publisher
 import org.reactivestreams.Subscriber
 import org.reactivestreams.Subscription
 
-private data class GraphQlSubscription(
-    /** Operation name (e.g., `"subscribeToMessages"`). */
-    val operation: String,
-    /** The [CloseReason] to send the client when the subscription is successfully completed. */
-    val completionReason: CloseReason,
-)
+private data class GraphQlSubscription(val path: String, val name: String, val isAuthenticated: Boolean)
+
+/** The [CloseReason] to send the client when the subscription is successfully completed. */
+private val completionReason = CloseReason(CloseReason.Codes.NORMAL, "The user deleted their account.")
 
 /** Adds routes to the [context] which deal with GraphQL subscriptions. */
 fun routeGraphQlSubscriptions(context: Routing) {
-    val completionReason = CloseReason(CloseReason.Codes.NORMAL, "The user deleted their account.")
-    val subscriptions = mapOf(
-        "messages-subscription" to "subscribeToMessages",
-        "chat-messages-subscription" to "subscribeToChatMessages",
-        "accounts-subscription" to "subscribeToAccounts",
-        "group-chats-subscription" to "subscribeToGroupChats",
-        "typing-statuses-subscription" to "subscribeToTypingStatuses",
-        "chat-typing-statuses-subscription" to "subscribeToChatTypingStatuses",
-        "online-statuses-subscription" to "subscribeToOnlineStatuses",
-        "chat-online-statuses-subscription" to "subscribeToChatOnlineStatuses",
-    )
-    for ((path, operation) in subscriptions)
-        routeSubscription(context, path, GraphQlSubscription(operation, completionReason))
+    setOf(
+        GraphQlSubscription(path = "messages-subscription", "subscribeToMessages", isAuthenticated = true),
+        GraphQlSubscription(path = "chat-messages-subscription", "subscribeToChatMessages", isAuthenticated = false),
+        GraphQlSubscription(path = "online-statuses-subscription", "subscribeToOnlineStatuses", isAuthenticated = true),
+        GraphQlSubscription(
+            path = "chat-online-statuses-subscription",
+            "subscribeToChatOnlineStatuses",
+            isAuthenticated = false,
+        ),
+        GraphQlSubscription(path = "typing-statuses-subscription", "subscribeToTypingStatuses", isAuthenticated = true),
+        GraphQlSubscription(
+            path = "chat-typing-statuses-subscription",
+            "subscribeToChatTypingStatuses",
+            isAuthenticated = false,
+        ),
+        GraphQlSubscription(path = "accounts-subscription", "subscribeToAccounts", isAuthenticated = true),
+        GraphQlSubscription(path = "chat-accounts-subscription", "subscribeToChatAccounts", isAuthenticated = false),
+        GraphQlSubscription(path = "group-chats-subscription", "subscribeToGroupChats", isAuthenticated = true),
+    ).forEach { routeSubscription(context, it) }
 }
 
 /**
- * Binds a WebSocket for the GraphQL [subscription] at the [path] (e.g., `"messages-subscription"`).
+ * Binds a WebSocket for the GraphQL [subscription].
  *
  * Once subscribed, the client will receive a [CreatedSubscription].
  *
@@ -59,14 +63,9 @@ fun routeGraphQlSubscriptions(context: Routing) {
  * with the [CloseReason.Codes.VIOLATED_POLICY]. If the subscription completes due to a server-side error, the
  * connection will be closed with the [CloseReason.Codes.INTERNAL_ERROR].
  */
-private fun routeSubscription(context: Routing, path: String, subscription: GraphQlSubscription): Unit = with(context) {
-    webSocket(path) {
-        val unauthenticatedOperations = setOf(
-            "subscribeToChatMessages",
-            "subscribeToChatOnlineStatuses",
-            "subscribeToChatTypingStatuses",
-        )
-        val token = if (subscription.operation in unauthenticatedOperations) null else incoming.receive() as Frame.Text
+private fun routeSubscription(context: Routing, subscription: GraphQlSubscription): Unit = with(context) {
+    webSocket(subscription.path) {
+        val token = if (subscription.isAuthenticated) incoming.receive() as Frame.Text else null
         try {
             val result = buildExecutionResult(this, token?.readText())
             if (result.errors.isEmpty()) subscribe(this, subscription, result) else closeWithError(this, result)
@@ -168,7 +167,7 @@ private class Subscriber(
     private fun sendCreatedSubscription() {
         val doc = mapOf(
             "data" to mapOf(
-                graphQlSubscription.operation to mapOf("__typename" to "CreatedSubscription", "placeholder" to ""),
+                graphQlSubscription.name to mapOf("__typename" to "CreatedSubscription", "placeholder" to ""),
             ),
         )
         val json = objectMapper.writeValueAsString(doc)
@@ -188,8 +187,8 @@ private class Subscriber(
         withSession { send(Frame.Text(json)) }
     }
 
-    /** Closes the connection with the [GraphQlSubscription.completionReason]. */
-    override fun onComplete(): Unit = withSession { close(graphQlSubscription.completionReason) }
+    /** Closes the connection with the [completionReason]. */
+    override fun onComplete(): Unit = withSession { close(completionReason) }
 
     /** Closes the connection with a [CloseReason.Codes.INTERNAL_ERROR]. */
     override fun onError(throwable: Throwable): Unit = withSession {